addressing comments: using repo config for connections, updating tests
